requirements.txt
alembic==1.15.2
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
cffi==1.17.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
ecdsa==0.19.1
et_xmlfile==2.0.0
fastapi==0.115.12
greenlet==3.2.1
h11==0.16.0
idna==3.10
lxml==5.4.0
Mako==1.3.10
MarkupSafe==3.0.2
numpy==2.2.5
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
passlib==1.7.4
pefile==2023.2.7
pillow==11.2.1
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.10.1
PyPDF2==3.0.1
PySide6==6.9.0
PySide6_Addons==6.9.0
PySide6_Essentials==6.9.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pytz==2025.2
pywin32==310
pywin32-ctypes==0.2.3
rsa==4.9.1
setuptools==80.3.1
shiboken6==6.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
watchdog==6.0.0
Whoosh==2.7.4
xlrd==2.0.1
